(PUP-7517) Allow agent daemon to time out stalled runs
On occasion, a puppet agent can end up waiting indefinitely on some process
that will never return or terminate. This patch adds a new setting, runtimeout,
that can be used to specify the maximum duration allowed for a puppet run. When
set to a non-zero value, the agent daemon will kill any agent run it starts
which exceeds this value.

This patch does not handle runs started outside of the agent daemon by
`puppet agent --onetime`.