(WIP)(PUP-7451) Add handling for Retry-After headers
Prior to this commit, the HTTP client used by the Puppet agent had special
handling for 3xx error codes in order to handle HTTP redirects. This patch
re-factors the response handler so that 503 and 429 error codes with
Retry-After headers set will result in a sleep followed by a retry of the
request. In particular, this allows a Puppet Server to re-splay its agent
population when overloaded by sending a 503 response with a randomized
Retry-After header, as implemented in SERVER-1767.