(WIP)(PUP-3647) Prune stale entries from the statefile
Puppet records the historical state of catalog resources in a statefile. This
information supports the audit and schedule metaparemeters which need
historical data to report on change or ensure that a resource is only synced a
given number of times in a schedule window. Currently, resources will remain in
the statefile indefinitely, even if they disappear from the catalog. This patch
adds a new `statettl` setting that is used to prune resources that the agent
has stopped interacting with because they are no longer present in catalogs.